requirements.txt
alembic==1.4.3
Babel==2.8.0
click==7.1.2
decorator==4.4.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
infinity==1.5
intervals==0.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2==2.8.6
python-dateutil==2.6.0
python-editor==1.0.4
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.19
validators==0.18.1
Werkzeug==1.0.1
WTForms==2.3.3
WTForms-Components==0.10.4
